JobManagement.get_job() causes FutureWarning
The python-dracclient job resource's JobManagement.get_job() causes a
FutureWarning. When that function is called, the following is output.

/usr/lib/python2.7/site-packages/dracclient/resources/job.py:103:
FutureWarning: The behavior of this method will change in future
versions. Use specific 'len(elem)' or 'elem is not None' test instead.
  if drac_job:

Change the identified line as suggested by the warning.

Change-Id: I9624b76cbf2e195f387ba104c98ae922619aa1f2
Closes-Bug: #1639339