Get/Set power state fails
It's possible for iDRACs to get into a state where the Name element is
not returned when querying DCIM_ComputerSystem.  This results in the
get_power_state command failing against nodes with iDRACs in this
state.

This patch removes selecting on the Name element from get_power_state.
This makes get_power_state work against both nodes with iDRACs in this
state and iDRACs not in this state.

Change-Id: Ide2d69b695376c777e5c4ee847dd0e481474a09f
Closes-Bug: #1667088