Increased is-iDRAC-ready timeout
The current is-iDRAC-ready timeout is set to 24 iterations times a
delay of 10 seconds, or 4 minutes.  This timeout is too short for 12G
and 13G hardware as operations against the iDRAC will frequently
timeout if the iDRAC is not in a ready state.

This patch increases the number of iterations to 48, which bumps the
timeout up to 8 minutes.  This increased timeout is sufficient for the
iDRAC on 14G, 13G, and 12G hardware to become ready and therefore avoid
a timeout exception.

Change-Id: I9dc476d3c28475919404077875f04fd006d4464a
Closes-Bug: 1718277