Allow invoke without checking the return value
Some versions of firmware for at least 13G hardware can return a
failure return value and a message indicating success on certain
invoke calls.  This patch modifies invoke so that it can be used to
execute the invoke command and return the result without raising an
exception if the return value indicates a failure.  This allows the
caller to inspect the returned message to determine success or failure.

Change-Id: I85a1d234d17ae601ed984109c8aa668aaf86eae2