Fix iDRAC reset
The current code waits for 3 consecutive failed pings, each 10 seconds
apart to determine that an iDRAC has gone down during an iDRAC reset.
This is too long for some servers, as the iDRAC may come back up before
the 3rd ping failure.  This results in a failure to detect the iDRAC
going down, which causes a timeout on the reset.

This patch changes the code to wait for only 2 consecutive ping
failures, which is what our highly tested downstream code does.

Closes-Bug: 1816195
Change-Id: Iac21d5eb722834089cbe4a2e7e19370a951951f0