Check LCStatus instead of MessageID to determine if an iDRAC is ready
We are currently checking the MessageID to determine if an iDRAC
is ready. This patch is to check the value of LCStatus instead of
MessageID. If the value of LCStatus is "0", then the iDRAC is
considered ready.

Change-Id: I3426c226c4bb3cdcc95b98e9b203f100bb6777d5
(cherry picked from commit bcce3bdc6b33c8d6f0721eac024ebff61de1e2db)